Update Ubuntu to latest LTS
The current container image contains a vulnerable version of `libwebp6`
(see [CVE-2023-4863]). While it shouldn't be possible to exploit the
vulnerability, the service is publicly exposed to the internet and we
should still patch it. Switching to the latest LTS of Ubuntu
automatically brings in a patched version of the WebP library.

[CVE-2023-4863]: https://nvd.nist.gov/vuln/detail/CVE-2023-4863